requirements.txt
numpy
Pillow
PyYAML
pytest
scikit-image
torch
torchvision
tqdm
wandb
lpips
